+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import HomePage from "./pages/Home";
+import HomePage from "./pages/HomePage";
 import './index.js';
 function App() {
   return (
