+++ src/components/Home/Listing/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Listing = () => {
+    return (
+        <div>Listing</div>
+    );
+}
+
+
+export default Listing;
