+++ src/components/Shared/index.js
@@ -1,0 +1,1 @@
+export { default as InputField } from '../Shared/InputField';
