+++ src/pages/HomePage/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Layout from '../../utils/Layout';
+import AddUser from "../../components/Home/AddUser";
+const HomePage = () => {
+    return (
+        <Layout>
+            <AddUser />
+        </Layout>
+    )
+}
+
+export default HomePage;
